requirements.txt
anyjson==0.3.3
bootstrap-py==0.8.2
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
docutils==0.15.2
et-xmlfile==1.0.1
future==0.18.2
get==2019.4.13
gevent==1.4.0
gitdb2==2.0.6
GitPython==3.0.4
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
imagesize==1.1.0
importlib-metadata==1.3.0
isort==4.3.21
jdcal==1.4.1
Jinja2==2.10.3
kiwisolver==1.1.0
kombu==4.6.7
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==8.0.2
openpyxl==2.6.3
packaging==19.2
pguard==0.4.0
Pillow==6.2.1
post==2019.4.13
product==0.1.1.1
psycopg2==2.8.4
public==2019.4.13
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
scipy==1.3.1
six==1.12.0
smmap2==2.0.5
snowballstemmer==2.0.0
sqlparse==0.3.0
transliterate==1.10.2
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
whitenoise==5.0.1
wrapt==1.11.2
zipp==0.6.0
